Fix dependency cycle between trap-caching and init
