+++ lib/codeql.js
@@ -23,7 +23,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getExtraOptions = exports.getCodeQLForCmd = exports.getCodeQLForTesting = exports.getCachedCodeQL = exports.setCodeQL = exports.getCodeQL = exports.setupCodeQL = exports.CODEQL_VERSION_RESOLVE_ENVIRONMENT = exports.CODEQL_VERSION_INIT_WITH_QLCONFIG = exports.CODEQL_VERSION_EXPORT_CODE_SCANNING_CONFIG = exports.CODEQL_VERSION_SECURITY_EXPERIMENTAL_SUITE = exports.CODEQL_VERSION_BETTER_RESOLVE_LANGUAGES = exports.CODEQL_VERSION_GHES_PACK_DOWNLOAD = exports.CommandInvocationError = void 0;
+exports.getTrapCachingExtractorConfigArgsForLang = exports.getTrapCachingExtractorConfigArgs = exports.getExtraOptions = exports.getCodeQLForCmd = exports.getCodeQLForTesting = exports.getCachedCodeQL = exports.setCodeQL = exports.getCodeQL = exports.setupCodeQL = exports.CODEQL_VERSION_RESOLVE_ENVIRONMENT = exports.CODEQL_VERSION_INIT_WITH_QLCONFIG = exports.CODEQL_VERSION_EXPORT_CODE_SCANNING_CONFIG = exports.CODEQL_VERSION_SECURITY_EXPERIMENTAL_SUITE = exports.CODEQL_VERSION_BETTER_RESOLVE_LANGUAGES = exports.CODEQL_VERSION_GHES_PACK_DOWNLOAD = exports.CommandInvocationError = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const core = __importStar(require("@actions/core"));
@@ -37,7 +37,6 @@ const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const setupCodeql = __importStar(require("./setup-codeql"));
 const toolrunner_error_catcher_1 = require("./toolrunner-error-catcher");
-const trap_caching_1 = require("./trap-caching");
 const util = __importStar(require("./util"));
 const util_1 = require("./util");
 class CommandInvocationError extends Error {
@@ -240,7 +239,7 @@ async function getCodeQLForCmd(cmd, checkVersion) {
             const extraArgs = config.languages.map((language) => `--language=${language}`);
             if (config.languages.filter((l) => (0, languages_1.isTracedLanguage)(l)).length > 0) {
                 extraArgs.push("--begin-tracing");
-                extraArgs.push(...(await (0, trap_caching_1.getTrapCachingExtractorConfigArgs)(config)));
+                extraArgs.push(...(await getTrapCachingExtractorConfigArgs(config)));
                 extraArgs.push(`--trace-process-name=${processName}`);
                 if (
                 // There's a bug in Lua tracing for Go on Windows in versions earlier than
@@ -316,7 +315,7 @@ async function getCodeQLForCmd(cmd, checkVersion) {
             await (0, toolrunner_error_catcher_1.toolrunnerErrorCatcher)(cmd, [
                 "database",
                 "trace-command",
-                ...(await (0, trap_caching_1.getTrapCachingExtractorConfigArgsForLang)(config, language)),
+                ...(await getTrapCachingExtractorConfigArgsForLang(config, language)),
                 ...getExtraOptionsFromEnv(["database", "trace-command"]),
                 databasePath,
                 "--",
@@ -813,4 +812,25 @@ async function getCodeScanningConfigExportArguments(config, codeql) {
     }
     return [];
 }
+// This constant sets the size of each TRAP cache in megabytes.
+const TRAP_CACHE_SIZE_MB = 1024;
+async function getTrapCachingExtractorConfigArgs(config) {
+    const result = [];
+    for (const language of config.languages)
+        result.push(await getTrapCachingExtractorConfigArgsForLang(config, language));
+    return result.flat();
+}
+exports.getTrapCachingExtractorConfigArgs = getTrapCachingExtractorConfigArgs;
+async function getTrapCachingExtractorConfigArgsForLang(config, language) {
+    const cacheDir = config.trapCaches[language];
+    if (cacheDir === undefined)
+        return [];
+    const write = await (0, actions_util_1.isAnalyzingDefaultBranch)();
+    return [
+        `-O=${language}.trap.cache.dir=${cacheDir}`,
+        `-O=${language}.trap.cache.bound=${TRAP_CACHE_SIZE_MB}`,
+        `-O=${language}.trap.cache.write=${write}`,
+    ];
+}
+exports.getTrapCachingExtractorConfigArgsForLang = getTrapCachingExtractorConfigArgsForLang;
 //# sourceMappingURL=codeql.js.map
