+++ lib/codeql.test.js
@@ -41,9 +41,9 @@ const actionsUtil = __importStar(require("./actions-util"));
 const codeql = __importStar(require("./codeql"));
 const defaults = __importStar(require("./defaults.json"));
 const feature_flags_1 = require("./feature-flags");
-const init_1 = require("./init");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
+const setup_codeql_1 = require("./setup-codeql");
 const testing_utils_1 = require("./testing-utils");
 const util = __importStar(require("./util"));
 const util_1 = require("./util");
@@ -125,7 +125,7 @@ function mockApiDetails(apiDetails) {
             const result = await codeql.setupCodeQL(url, testing_utils_1.SAMPLE_DOTCOM_API_DETAILS, tmpDir, util.GitHubVariant.DOTCOM, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
             t.assert(toolcache.find("CodeQL", `0.0.0-${version}`));
             t.is(result.toolsVersion, `0.0.0-${version}`);
-            t.is(result.toolsSource, init_1.ToolsSource.Download);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
             t.assert(Number.isInteger(result.toolsDownloadDurationMs));
         }
         t.is(toolcache.findAllVersions("CodeQL").length, 2);
@@ -145,7 +145,7 @@ function mockApiDetails(apiDetails) {
         const result = await codeql.setupCodeQL(url, testing_utils_1.SAMPLE_DOTCOM_API_DETAILS, tmpDir, util.GitHubVariant.DOTCOM, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
         t.assert(toolcache.find("CodeQL", "0.0.0-20200610"));
         t.deepEqual(result.toolsVersion, "0.0.0-20200610");
-        t.is(result.toolsSource, init_1.ToolsSource.Download);
+        t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
         t.assert(Number.isInteger(result.toolsDownloadDurationMs));
     });
 });
@@ -180,7 +180,7 @@ for (const { cliVersion, expectedToolcacheVersion, } of EXPLICITLY_REQUESTED_BUN
             t.assert(releaseApiMock.isDone(), "Releases API should have been called");
             t.assert(toolcache.find("CodeQL", expectedToolcacheVersion));
             t.deepEqual(result.toolsVersion, cliVersion);
-            t.is(result.toolsSource, init_1.ToolsSource.Download);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
             t.assert(Number.isInteger(result.toolsDownloadDurationMs));
         });
     });
@@ -202,7 +202,7 @@ for (const toolcacheVersion of [
             sinon.stub(toolcache, "findAllVersions").returns([toolcacheVersion]);
             const result = await codeql.setupCodeQL(undefined, testing_utils_1.SAMPLE_DOTCOM_API_DETAILS, tmpDir, util.GitHubVariant.DOTCOM, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
             t.is(result.toolsVersion, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION.cliVersion);
-            t.is(result.toolsSource, init_1.ToolsSource.Toolcache);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Toolcache);
             t.is(result.toolsDownloadDurationMs, undefined);
         });
     });
@@ -221,7 +221,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
                 tagName: defaults.bundleVersion,
             }, (0, logging_1.getRunnerLogger)(true), false);
             t.deepEqual(result.toolsVersion, "0.0.0-20200601");
-            t.is(result.toolsSource, init_1.ToolsSource.Toolcache);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Toolcache);
             t.is(result.toolsDownloadDurationMs, undefined);
             const cachedVersions = toolcache.findAllVersions("CodeQL");
             t.is(cachedVersions.length, 1);
@@ -243,7 +243,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
                 tagName: defaults.bundleVersion,
             }, (0, logging_1.getRunnerLogger)(true), false);
             t.deepEqual(result.toolsVersion, defaults.cliVersion);
-            t.is(result.toolsSource, init_1.ToolsSource.Download);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
             t.assert(Number.isInteger(result.toolsDownloadDurationMs));
             const cachedVersions = toolcache.findAllVersions("CodeQL");
             t.is(cachedVersions.length, 2);
@@ -263,7 +263,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
         });
         const result = await codeql.setupCodeQL("latest", testing_utils_1.SAMPLE_DOTCOM_API_DETAILS, tmpDir, util.GitHubVariant.DOTCOM, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
         t.deepEqual(result.toolsVersion, defaults.cliVersion);
-        t.is(result.toolsSource, init_1.ToolsSource.Download);
+        t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
         t.assert(Number.isInteger(result.toolsDownloadDurationMs));
         const cachedVersions = toolcache.findAllVersions("CodeQL");
         t.is(cachedVersions.length, 2);
@@ -306,7 +306,7 @@ for (const isBundleVersionInUrl of [true, false]) {
                 cliVersion: defaults.cliVersion,
                 tagName: defaults.bundleVersion,
             }, (0, logging_1.getRunnerLogger)(true), false);
-            t.is(result.toolsSource, init_1.ToolsSource.Download);
+            t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
             t.assert(Number.isInteger(result.toolsDownloadDurationMs));
             const cachedVersions = toolcache.findAllVersions("CodeQL");
             t.is(cachedVersions.length, 1);
@@ -329,7 +329,7 @@ for (const isBundleVersionInUrl of [true, false]) {
         });
         const result = await codeql.setupCodeQL("https://github.com/codeql-testing/codeql-cli-nightlies/releases/download/codeql-bundle-20230203/codeql-bundle.tar.gz", testing_utils_1.SAMPLE_DOTCOM_API_DETAILS, tmpDir, util.GitHubVariant.DOTCOM, testing_utils_1.SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
         t.is(result.toolsVersion, "0.0.0-20230203");
-        t.is(result.toolsSource, init_1.ToolsSource.Download);
+        t.is(result.toolsSource, setup_codeql_1.ToolsSource.Download);
         t.true(Number.isInteger(result.toolsDownloadDurationMs));
         const cachedVersions = toolcache.findAllVersions("CodeQL");
         t.is(cachedVersions.length, 1);
