+++ lib/init-action.js
@@ -35,6 +35,7 @@ const init_1 = require("./init");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
+const setup_codeql_1 = require("./setup-codeql");
 const trap_caching_1 = require("./trap-caching");
 const util_1 = require("./util");
 const workflow_1 = require("./workflow");
@@ -45,7 +46,7 @@ async function sendCompletedStatusReport(startedAt, config, toolsDownloadDuratio
         ...statusReportBase,
         tools_input: (0, actions_util_1.getOptionalInput)("tools") || "",
         tools_resolved_version: toolsVersion,
-        tools_source: toolsSource || init_1.ToolsSource.Unknown,
+        tools_source: toolsSource || setup_codeql_1.ToolsSource.Unknown,
         workflow_languages: workflowLanguages || "",
     };
     const initToolsDownloadFields = {};
