+++ lib/init.js
@@ -23,7 +23,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.installPythonDeps = exports.runInit = exports.initConfig = exports.initCodeQL = exports.ToolsSource = void 0;
+exports.installPythonDeps = exports.runInit = exports.initConfig = exports.initCodeQL = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
@@ -34,13 +34,6 @@ const configUtils = __importStar(require("./config-utils"));
 const feature_flags_1 = require("./feature-flags");
 const tracer_config_1 = require("./tracer-config");
 const util = __importStar(require("./util"));
-var ToolsSource;
-(function (ToolsSource) {
-    ToolsSource["Unknown"] = "UNKNOWN";
-    ToolsSource["Local"] = "LOCAL";
-    ToolsSource["Toolcache"] = "TOOLCACHE";
-    ToolsSource["Download"] = "DOWNLOAD";
-})(ToolsSource || (exports.ToolsSource = ToolsSource = {}));
 async function initCodeQL(toolsInput, apiDetails, tempDir, variant, defaultCliVersion, logger) {
     logger.startGroup("Setup CodeQL tools");
     const { codeql, toolsDownloadDurationMs, toolsSource, toolsVersion } = await (0, codeql_1.setupCodeQL)(toolsInput, apiDetails, tempDir, variant, defaultCliVersion, logger, true);
