+++ lib/trap-caching.js
@@ -23,7 +23,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getTotalCacheSize = exports.getLanguagesSupportingCaching = exports.uploadTrapCaches = exports.downloadTrapCaches = exports.getTrapCachingExtractorConfigArgsForLang = exports.getTrapCachingExtractorConfigArgs = void 0;
+exports.getTotalCacheSize = exports.getLanguagesSupportingCaching = exports.uploadTrapCaches = exports.downloadTrapCaches = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const cache = __importStar(require("@actions/cache"));
@@ -36,8 +36,6 @@ const util_1 = require("./util");
 // this for CLI/extractor changes, since the CLI version also
 // goes into the cache key.
 const CACHE_VERSION = 1;
-// This constant sets the size of each TRAP cache in megabytes.
-const CACHE_SIZE_MB = 1024;
 // This constant sets the minimum size in megabytes of a TRAP
 // cache for us to consider it worth uploading.
 const MINIMUM_CACHE_MB_TO_UPLOAD = 10;
@@ -46,25 +44,6 @@ const MINIMUM_CACHE_MB_TO_UPLOAD = 10;
 // this timeout is per operation, so will be run as many
 // times as there are languages with TRAP caching enabled.
 const MAX_CACHE_OPERATION_MS = 120000; // Two minutes
-async function getTrapCachingExtractorConfigArgs(config) {
-    const result = [];
-    for (const language of config.languages)
-        result.push(await getTrapCachingExtractorConfigArgsForLang(config, language));
-    return result.flat();
-}
-exports.getTrapCachingExtractorConfigArgs = getTrapCachingExtractorConfigArgs;
-async function getTrapCachingExtractorConfigArgsForLang(config, language) {
-    const cacheDir = config.trapCaches[language];
-    if (cacheDir === undefined)
-        return [];
-    const write = await actionsUtil.isAnalyzingDefaultBranch();
-    return [
-        `-O=${language}.trap.cache.dir=${cacheDir}`,
-        `-O=${language}.trap.cache.bound=${CACHE_SIZE_MB}`,
-        `-O=${language}.trap.cache.write=${write}`,
-    ];
-}
-exports.getTrapCachingExtractorConfigArgsForLang = getTrapCachingExtractorConfigArgsForLang;
 /**
  * Download TRAP caches from the Actions cache.
  * @param codeql The CodeQL instance to use.
