+++ lib/trap-caching.test.js
@@ -135,7 +135,7 @@ function getTestConfigWithTempDir(tmpDir) {
     await util.withTmpDir(async (tmpDir) => {
         const config = getTestConfigWithTempDir(tmpDir);
         sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
-        const result = await (0, trap_caching_1.getTrapCachingExtractorConfigArgsForLang)(config, languages_1.Language.javascript);
+        const result = await (0, codeql_1.getTrapCachingExtractorConfigArgsForLang)(config, languages_1.Language.javascript);
         t.deepEqual(result, [
             `-O=javascript.trap.cache.dir=${path.resolve(tmpDir, "jsCache")}`,
             "-O=javascript.trap.cache.bound=1024",
@@ -147,7 +147,7 @@ function getTestConfigWithTempDir(tmpDir) {
     await util.withTmpDir(async (tmpDir) => {
         const config = getTestConfigWithTempDir(tmpDir);
         sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(false);
-        const result = await (0, trap_caching_1.getTrapCachingExtractorConfigArgs)(config);
+        const result = await (0, codeql_1.getTrapCachingExtractorConfigArgs)(config);
         t.deepEqual(result, [
             `-O=javascript.trap.cache.dir=${path.resolve(tmpDir, "jsCache")}`,
             "-O=javascript.trap.cache.bound=1024",
